requirements.txt
# Single file for Streamlit Cloud (no -r to avoid install issues)
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
google-generativeai>=0.8.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
email-validator>=2.0.0
selenium>=4.0.0
webdriver-manager>=4.0.0
